Add a upg_uri_copy function
I'm writing something else right now, and the ability to copy a URI is
helpful.

Another way to implement this would be to_string() then _new(). I was
worried the to_string would be too expensive, but it frankly might be
worth it given the amount of copies we do.